requirements.txt
flask
flask_sqlalchemy
python-dotenv
sqlalchemy
flask-migrate
flask-cors
gunicorn